pyproject.toml
[project]
name = "ds340-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "llvmlite>=0.43.0",
    "matplotlib>=3.9.2",
    "numba>=0.60.0",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "polars>=1.14.0",
    "pyarrow>=18.0.0",
    "pymatting>=1.1.13",
    "rembg>=2.0.59",
    "ruff>=0.7.4",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
]

[project.optional-dependencies]
gpu = [
    "rembg[gpu]>=2.0.59",
]

[tool.pyright]
include = ["src"]

[tool.ruff]
include = ["pyproject.toml"]
line-length = 125
select = ["ALL"]
ignore = ["G004", "D211", "D213", "UP012", "T201"]
